video_processor/pyproject.toml
[project]
name = "ballskicker-video-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "boto3>=1.36.1",
    "ffmpeg-python>=0.2.0",
    "opencv-python>=4.11.0.86",
    "psutil>=7.0.0",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "typer>=0.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["video_processor"]

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "pre-commit>=4.1.0",
    "ruff>=0.9.2",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
